+++ packages/dara-core/js/types/core.tsx
@@ -81,19 +81,20 @@ export interface BackendStore extends PersistenceStore {
     readonly: boolean;
 }
 
-export interface SingleVariable<T = any, TStore extends PersistenceStore = never> {
+export interface QueryParamStore extends PersistenceStore {
+    __typename: 'QueryParamStore';
+    query: string;
+}
+
+export interface BrowserStore extends PersistenceStore {
+    __typename: 'BrowserStore';
+}
+
+export interface SingleVariable<T = any, TStore extends PersistenceStore = PersistenceStore> {
     __typename: 'Variable';
     default: T | DerivedVariable;
     nested: string[];
-    persist_value?: boolean;
     store?: TStore;
-    uid: string;
-}
-
-export interface UrlVariable<T> {
-    __typename: 'UrlVariable';
-    default: T;
-    query: string;
     uid: string;
 }
 
@@ -145,12 +146,11 @@ export interface ServerVariable {
 
 export type AnyVariable<T> =
     | SingleVariable<T>
-    | UrlVariable<T>
     | DerivedVariable
     | SwitchVariable<T>
     | StateVariable
     | ServerVariable;
-export type Variable<T> = SingleVariable<T> | UrlVariable<T> | DerivedVariable | SwitchVariable<T> | StateVariable;
+export type Variable<T> = SingleVariable<T> | DerivedVariable | SwitchVariable<T> | StateVariable;
 
 export interface ResolvedDerivedVariable {
     deps: Array<number>;
@@ -351,7 +351,6 @@ export interface DaraEventMap {
     SERVER_COMPONENT_LOADED: { name: string; uid: string; value: ComponentInstance };
     DERIVED_VARIABLE_LOADED: { variable: DerivedVariable; value: any };
     PLAIN_VARIABLE_LOADED: { variable: SingleVariable<any>; value: any };
-    URL_VARIABLE_LOADED: { variable: UrlVariable<any>; value: any };
     STATE_VARIABLE_LOADED: { variable: StateVariable; value: any };
 }
 
@@ -402,7 +401,7 @@ export interface ActionImpl {
 }
 
 export interface UpdateVariableImpl extends ActionImpl {
-    variable: SingleVariable<any> | UrlVariable<any>;
+    variable: SingleVariable<any>;
     value: any;
 }